requirements.txt
astropy==6.0.1
astropy-iers-data==0.2024.6.17.0.31.35
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyerfa==2.0.1.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tzdata==2024.1
zipp==3.19.2
